src: Switch complexMatrix to boost.
Switch the data container class for the complexMatrix from root's
TMatrixT<double> to boost::numeric::ublas::matrix. Like this, one can calculate
the determinant of complex matrices quite easily.

The version number in the ClassDef had to be raised for the complexMatrix
class.

+++ src/fitResult.cc
@@ -699,8 +699,8 @@ fitResult::fill
 		cout << "fitResult::fill(): warning: number of production amplitudes "
 		     << "(" << _prodAmps.size() << ") does not match number of "
 		     << "covariance matrix indices (" << _fitParCovMatrixIndices.size() << ")." << endl;
-	if (   ((int)_waveNames.size() != _normIntegral.nRows())
-	    or ((int)_waveNames.size() != _normIntegral.nCols()))
+	if (   (_waveNames.size() != _normIntegral.nRows())
+	    or (_waveNames.size() != _normIntegral.nCols()))
 		cout << "fitResult::fill(): warning: number of waves (" << _waveNames.size()
 		     << ") does not match size of normalization integral "
 		     << "(" << _normIntegral.nRows() << ", " << _normIntegral.nCols() << ")." << endl;
